[*] pierdolę.

+++ Rasterizer/Rasterizer/src/Math.h
@@ -4,7 +4,7 @@
 #include <math.h>
 #include <Windows.h>
 
-#define piDiv180 (M_PI / 180.0f)
+#define piDiv180 ((float)M_PI / 180.0f)
 
 #define Clamp(x, a, b) min(max((x),(a)), (b))
 #define FloatLerp(x, y, s) (x)*(1.0f - (s)) + (y)*(s)
